test: refactor adaptive throttler and monitor suites
